services/embedding-service/requirements.txt
fastapi>=0.85.1,<0.96.0
uvicorn==0.22.0
pydantic>=1.10.0,<2.0.0
boto3==1.26.90
psycopg2-binary==2.9.6
chromadb-client==0.5.6.dev0
openai==0.27.6
requests==2.30.0
python-dotenv==1.0.0